Added song route, converted buttons to component
Added route for a guess the song page in addition to album and rapper. Also converted buttons to be a component. I want to try and make use of components as much as possible on this project, and i think that stuff like this is a good start to getting that goal started