Creation of basic search method
Added react select package to give dropdown that gives an easy way to search through a list to choose an option. Just basic testing, but this seems to be the simplest implementation of a search method.
Will need to work out other logic for sending chosen to logic that checks if it is correct, as well as potentially removing guessed options from the pool of answers once guessed